feat(app): order details
