+++ packages/app/src/components/UserVaultsContainer/index.ts
@@ -1,0 +1,1 @@
+export * from './UserOrdersContainer';
